refactor: remove redundant code, improve authentication, and enhance type safety, bug fix
- Fixed an issue where the comment count was not updated after a comment was deleted.
- Removed unused `author` field in `AddBlogSchema` and refactored blog creation to use `verifyToken` for author validation.
- Added server-side validation for missing blog IDs in comment handling.
- Cleaned up unused imports and console logs across multiple files.
- Enhanced consistency in code formatting and alignment throughout the project.

+++ app/api/comment/[id]/route.ts
@@ -1,5 +1,6 @@
 import connectDb from "@/lib/db";
 import { ErrorHandler } from "@/lib/errorHandler";
+import Blog from "@/models/blog.model";
 import Comment from "@/models/comment.model";
 import { NextRequest, NextResponse } from "next/server";
 
@@ -10,10 +11,20 @@ export async function DELETE(
   try {
     await connectDb();
     const { id } = await params;
+    const { searchParams } = new URL(req.url);
+
+    const blogId = searchParams.get("blogId");
+
+    if (!blogId) return ErrorHandler("Something went wrong");
 
     if (!id) return ErrorHandler("Comment id required", 400);
 
     const comment = await Comment.findByIdAndDelete(id);
+    if (!comment) return ErrorHandler("Comment not found", 404);
+
+    await Blog.findByIdAndUpdate(blogId, {
+      $pull: { comments: comment._id },
+    });
 
     if (!comment) return ErrorHandler("Comment not found", 404);
